Minor changes according to remarks.

+++ src/main/java/ua/softserve/rv_018/greentourism/repository/EventRepository.java
@@ -9,8 +9,7 @@ import org.springframework.stereotype.Repository;
 import ua.softserve.rv_018.greentourism.model.Event;
 
 @Repository
-public interface EventRepository extends JpaRepository<Event, Long> {
-	
+public interface EventRepository extends JpaRepository<Event, Integer> {
 	@Query(value="select event.* "
 			+ "from place inner join point on event.point_id = point.id "
 			+ "where point.latitude >= :south_west_latitude "
